split api's up
